other minor changes
